# public/openapi.yaml
openapi: 3.1.0
info:
  title: Mayfly Orchestrator API
  version: "1.0.0"
  description: |
    The control-plane API for provisioning disposable, per-team CTF challenge instances on
    Vercel Sandbox. This is the Vercel side only — CTFd, its MySQL, and its Redis stay on
    Google Cloud and are never touched directly by this API (see the project's CLAUDE.md,
    "the boundary").

    **Two callers, two trust levels:**
    - The **CTFd plugin** (`ctfd_mayfly`) calls `POST /api/instances` and the submission
      webhook with an HMAC-SHA256 signature over the raw request body — see the
      `MayflySignature` security scheme below.
    - The **demo console** (`/`, a headless test harness standing in for CTFd — see
      CLAUDE.md, "There is no player UI in this repo") calls the unsigned `POST /api/launch`
      equivalent directly from the browser. There's no secret a browser could hold that
      wouldn't also be readable by anyone who opens devtools, so that boundary simply
      doesn't apply to it the way it does to CTFd.

    Every launch call returns a `runId` and a `runToken` immediately, before the sandbox
    exists — CTFd's own gunicorn worker pool is bounded, and blocking would exhaust it during
    an event's opening rush. Poll `GET /api/instances/{runId}` for the actual lifecycle,
    presenting `runToken` on every call — see `RunToken` below.

    **Auth summary, endpoint by endpoint**:
...
      contract.
...
      genuinely unforgeable without `MAYFLY_SIGNING_SECRET`.
...
servers:
...
components:
  securitySchemes:
    MayflySignature:
      type: apiKey
      in: header
      name: x-mayfly-signature
      description: |
        HMAC-SHA256 over `${timestampMs}.${rawRequestBody}` using the shared
        `MAYFLY_SIGNING_SECRET`, hex-encoded. Verified against the RAW body bytes (via
        `request.text()`), never a re-serialized object — `JSON.stringify` and Python's
        `json.dumps` disagree on whitespace and key order, so signing anything but the
        original bytes would make verification flaky. Must be sent alongside the
        `x-mayfly-timestamp` header (see below); requests older than a short clock-skew
        window are rejected even with a valid signature.
    RunToken:
      type: apiKey
      in: header
      name: x-mayfly-run-token
      description: |
        The opaque `runToken` returned by whichever launch call (`POST /api/instances` or
        `POST /api/launch`) created this run. Minted server-side (HMAC over the runId plus an
        expiry, using a secret distinct from `MAYFLY_SIGNING_SECRET` — see
        `lib/auth/run-token.ts`) and never computed by any caller, including CTFd's plugin,
        which only stores and echoes it back. Required on every
        `GET/POST /api/instances/{runId}[/extend|/stop]` call for that runId.
...
  schemas:
...
    LaunchRequest:
      type: object
      required: [challengeId, teamId, ttlSeconds]
      properties:
        challengeId:
          type: string
          description: >
            Either a demo-console fixture id (see lib/fixtures/challenges.ts) or an arbitrary
            caller-owned id — CTFd's plugin passes its own numeric challenge id and supplies
            image/port/vcpus directly instead, so this never needs to match a fixture.
          example: "file-explorer"
        teamId:
          type: string
          example: "team-42"
        ttlSeconds:
          type: integer
          minimum: 1
          maximum: 86400
          description: Active lifetime before the "expiring" grace window begins.
          example: 1800
        image:
          type: string
          description: A Vercel Container Registry reference. Present when the caller (typically CTFd) already owns its own challenge config instead of relying on a fixture lookup.
          example: "file-explorer:latest"
        port:
          type: integer
          description: Required when `image` is present.
          example: 3001
        vcpus:
          type: integer
          example: 1
        startCommand:
          type: string
          description: Overrides the image's own `/start.sh` convention. Rarely needed.
    LaunchResponse:
      type: object
      required: [runId, runToken]
      properties:
        runId:
          type: string
          example: "wrun_01M361RFZ35EHZFEDAF2PMNR79"
        runToken:
          type: string
          description: Present this on every later status/extend/stop call for this runId — see the `RunToken` security scheme.
          example: "1758620400000.9f2c7a1e4b6d8035..."
...
    Error:
      type: object
      properties:
        error:
          description: Either a plain message string, or a Zod flatten() validation error object.
paths:
  /api/instances:
    post:
      tags: [Launch]
      summary: Launch an instance (CTFd's signed contract)
      description: |
        The CTFd plugin's launch endpoint — every request must be HMAC-signed (see the
        `MayflySignature` security scheme). Returns a `runId` immediately; the actual
        provisioning happens asynchronously in a durable Workflow run. Idempotent from the
        caller's side only in the sense that CTFd's own plugin resolves "already have a live
        instance for this owner+challenge" before ever calling this — the orchestrator
        itself always starts a fresh run per call.
      security:
        - MayflySignature: []
      parameters:
        - name: x-mayfly-timestamp
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LaunchRequest"
      responses:
        "200":
          description: Launch accepted; a Workflow run has started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaunchResponse"
        "400":
          description: Schema validation failed, or challengeId matched no fixture and no image was supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid HMAC signature, or the timestamp is outside the allowed clock-skew window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/launch:
    post:
      tags: [Launch]
      summary: Launch an instance (unsigned, browser-facing demo console)
      description: |
        Same request/response contract as `POST /api/instances`, but unsigned — same-origin
        browser calls from the demo console (`/`) have no secret they could hold without
        leaking it to every visitor, so the HMAC boundary doesn't apply here. Never call this
        from CTFd; use the signed endpoint instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LaunchRequest"
      responses:
        "200":
          description: Launch accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaunchResponse"
        "400":
          description: Schema validation failed, or an unknown challengeId with no image supplied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...